+++ src/main/java/tetrad/Mutil.java
@@ -315,86 +315,6 @@ public class Mutil {
         System.out.print("\033[0G");  // Move cursor back to the beginning of the line
     }
     /**
-     * Prints the game header
-     * @deprecated WILL BE MOVED TO TERMINALGAME
-     */
-    @Deprecated
-    public static void printHeader() {
-        // title card
-        String title = "Terminal Trader";
-        String line1 = "-".repeat(MENU_WIDTH);
-        String line2 = "~ ".repeat((MENU_WIDTH - title.length())/4) + title + " ~".repeat((MENU_WIDTH - title.length())/4) + " ";
-        String line3 = "-".repeat(MENU_WIDTH);
-
-        switch (Game.headerSetting) {
-            case 0 -> {
-                line1 = blue(line1);
-                line2 = blue(line2);
-                line3 = blue(line3);
-            }
-            case 1 -> {
-                line1 = green(line1);
-                line2 = green(line2);
-                line3 = green(line3);
-            }
-            case 2 -> {
-                line1 = magenta(line1);
-                line2 = magenta(line2);
-                line3 = magenta(line3);
-            }
-            case 3 -> {
-                line1 = cyan(line1);
-                line2 = cyan(line2);
-                line3 = cyan(line3);
-            }
-            case 4 -> {
-                line1 = yellow(line1);
-                line2 = yellow(line2);
-                line3 = yellow(line3);
-            }
-            case 5 -> {
-                line1 = yellowB(line1);
-                line2 = yellowB(line2);
-                line3 = yellowB(line3);
-            }
-            case 6 -> {
-                line1 = greenB(line1);
-                line2 = greenB(line2);
-                line3 = greenB(line3);
-            }
-            case 7 -> {
-                line1 = redB(line1);
-                line2 = redB(line2);
-                line3 = redB(line3);
-            }
-            case 8 -> {
-                line1 = cyanB(line1);
-                line2 = cyanB(line2);
-                line3 = cyanB(line3);
-            }
-            case 9 -> {
-                line1 = red(line1);
-                line2 = red(line2);
-                line3 = red(line3);
-            }
-            case 10 -> {
-                line1 = magentaB(line1);
-                line2 = magentaB(line2);
-                line3 = magentaB(line3);
-            }
-            case 11 -> {
-                line1 = blueB(line1);
-                line2 = blueB(line2);
-                line3 = blueB(line3);
-            }
-        }
-
-        // print header
-        System.out.println(line1);
-        System.out.println(line2);
-        System.out.println(line3);
-    }
-    /**
      * Experimental function for printing taxman screen from ASCI art
      * @deprecated WILL BE MOVED TO TERMINALGAME
      */
