+++ src/main/java/tetrad/TerminalGame.java
@@ -13,20 +13,25 @@ import java.util.Scanner;
 
 import static tetrad.Mutil.MENU_WIDTH;
 import static tetrad.Mutil.blue;
+import static tetrad.Mutil.blueB;
 import static tetrad.Mutil.center;
 import static tetrad.Mutil.clearLine;
 import static tetrad.Mutil.clearScreen;
 import static tetrad.Mutil.cyan;
+import static tetrad.Mutil.cyanB;
 import static tetrad.Mutil.dollar;
 import static tetrad.Mutil.green;
+import static tetrad.Mutil.greenB;
 import static tetrad.Mutil.italic;
+import static tetrad.Mutil.magenta;
+import static tetrad.Mutil.magentaB;
 import static tetrad.Mutil.numColor;
 import static tetrad.Mutil.pause;
-import static tetrad.Mutil.printHeader;
 import static tetrad.Mutil.printTaxMan;
 import static tetrad.Mutil.red;
 import static tetrad.Mutil.redB;
 import static tetrad.Mutil.yellow;
+import static tetrad.Mutil.yellowB;
 
 public class TerminalGame extends Game {
     public TerminalGame() {
@@ -515,5 +520,85 @@ public class TerminalGame extends Game {
                 default -> System.out.println("Invalid Input");
             }   
         }
+    }
+
+    // static menu functions
+    /**
+     * Prints the game header
+     */
+    public static void printHeader() {
+        // title card
+        String title = "Terminal Trader";
+        String line1 = "-".repeat(MENU_WIDTH);
+        String line2 = "~ ".repeat((MENU_WIDTH - title.length())/4) + title + " ~".repeat((MENU_WIDTH - title.length())/4) + " ";
+        String line3 = "-".repeat(MENU_WIDTH);
+
+        switch (headerSetting) {
+            case 0 -> {
+                line1 = blue(line1);
+                line2 = blue(line2);
+                line3 = blue(line3);
+            }
+            case 1 -> {
+                line1 = green(line1);
+                line2 = green(line2);
+                line3 = green(line3);
+            }
+            case 2 -> {
+                line1 = magenta(line1);
+                line2 = magenta(line2);
+                line3 = magenta(line3);
+            }
+            case 3 -> {
+                line1 = cyan(line1);
+                line2 = cyan(line2);
+                line3 = cyan(line3);
+            }
+            case 4 -> {
+                line1 = yellow(line1);
+                line2 = yellow(line2);
+                line3 = yellow(line3);
+            }
+            case 5 -> {
+                line1 = yellowB(line1);
+                line2 = yellowB(line2);
+                line3 = yellowB(line3);
+            }
+            case 6 -> {
+                line1 = greenB(line1);
+                line2 = greenB(line2);
+                line3 = greenB(line3);
+            }
+            case 7 -> {
+                line1 = redB(line1);
+                line2 = redB(line2);
+                line3 = redB(line3);
+            }
+            case 8 -> {
+                line1 = cyanB(line1);
+                line2 = cyanB(line2);
+                line3 = cyanB(line3);
+            }
+            case 9 -> {
+                line1 = red(line1);
+                line2 = red(line2);
+                line3 = red(line3);
+            }
+            case 10 -> {
+                line1 = magentaB(line1);
+                line2 = magentaB(line2);
+                line3 = magentaB(line3);
+            }
+            case 11 -> {
+                line1 = blueB(line1);
+                line2 = blueB(line2);
+                line3 = blueB(line3);
+            }
+        }
+
+        // print header
+        System.out.println(line1);
+        System.out.println(line2);
+        System.out.println(line3);
     }
 }
